crash API uploaded

+++ Code/acts.py
@@ -24,7 +24,7 @@ sess = Session()
 app.config["MONGO_URI"] = "mongodb://localhost:27017/cc_db"
 mongo = PyMongo(app)
 
-
+crash = false;
 
 # Functions
 
@@ -65,6 +65,8 @@ class Cats(Resource):
 
 	def get(self, cname=None):
 		incrementHTTP()
+		if(crash==true):
+			return "", 500
 		if(cname):
 			return "", 405
 		cursor = mongo.db.cats.find({}, {"_id": 0, "update_time": 0})
@@ -76,10 +78,12 @@ class Cats(Resource):
 				return "There are no Categories", 204
 			else:
 				data = mergeDicts(data)
-				return data, 200
+				return data, 200		
 
 	def post(self, cname=None):
 		incrementHTTP()
+		if(crash==true):
+			return "", 500
 		if(cname):
 			return "", 405
 		data = request.get_json()
@@ -99,6 +103,8 @@ class Cats(Resource):
 
 	def delete(self,cname=None):
 		incrementHTTP()
+		if(crash==true):
+			return "", 500
 		if(cname):
 			x = mongo.db.cats.distinct(cname)
 			if(x==[]):
@@ -117,7 +123,8 @@ class Acts(Resource):
 
 	def get(self, cname=None, actId=None):
 		incrementHTTP()
-
+		if(crash==true):
+			return "", 500
 		if(actId):
 			return "",405
 
@@ -179,7 +186,8 @@ class Acts(Resource):
 
 	def post(self, cname=None, actId=None):
 		incrementHTTP()
-
+		if(crash==true):
+			return "", 500
 		if(actId):
 			return "",405
 
@@ -243,6 +251,8 @@ class Acts(Resource):
 
 	def delete(self,actId=None,cname=None):
 		incrementHTTP()
+		if(crash==true):
+			return "", 500
 		if(cname):
 			return "",405
 		if(not(actId)):
@@ -266,6 +276,8 @@ class numberActs(Resource):
 
 	def get(self,cname):
 		incrementHTTP()
+		if(crash==true):
+			return "", 500
 		temp = mongo.db.cats.distinct(cname)
 		if(temp == []):
 			return "Category Name does not exist", 405
@@ -282,6 +294,8 @@ class upvote(Resource):
 
 	def post(self):
 		incrementHTTP()
+		if(crash==true):
+			return "", 500
 		data = request.get_json()
 		try:
 			actId = data[0]
@@ -303,6 +317,8 @@ class actCount(Resource):
 
 	def get(self):
 		incrementHTTP()
+		if(crash==true):
+			return "", 500
 		count = mongo.db.acts.count()
 		if(count>=0):
 			return [count], 200
@@ -312,6 +328,8 @@ class actCount(Resource):
 class httpCount(Resource):
 
 	def get(self):
+		if(crash==true):
+			return "", 500
 		count = mongo.db.http.count()
 		if(count>=0):
 			return [count], 200
@@ -319,8 +337,18 @@ class httpCount(Resource):
 			return "Method not Allowed", 405
 
 	def delete(self):
+		if(crash==true):
+			return "", 500
 		r = mongo.db.http.drop()
 		return {}, 200
+
+class health(Resource):
+
+	def post(self):
+		if(crash==true):
+			return "", 500
+		crash = true;
+		return "", 200
 
 # Resources for Categories
 
@@ -343,6 +371,8 @@ api.add_resource(actCount,"/api/v1/acts/count",endpoint="actcount")
 
 api.add_resource(httpCount,"/api/v1/_count",endpoint="httpcount")
 
+# Resources for Health
+api.add_resource(health,"/api/v1/_crash",endpoint="crash")
 
 # Run the App
 if __name__ == "__main__":
